+++ pynapple/core/_jitted_functions.py
@@ -398,105 +398,6 @@ def pjitconvolve(data_array, array, trim="both"):
     return new_data_array
 
 
-@njit(parallel=True)
-def jitcontinuous_perievent(
-    time_array, data_array, time_target_array, starts, ends, windowsize
-):
-    N_samples = len(time_array)
-    N_target = len(time_target_array)
-    N_epochs = len(starts)
-    count = np.zeros((N_epochs, 2), dtype=np.int64)
-    start_t = np.zeros((N_epochs, 2), dtype=np.int64)
-
-    k = 0  # Epochs
-    t = 0  # Samples
-    i = 0  # Target
-
-    while ends[k] < time_array[t] and ends[k] < time_target_array[i]:
-        k += 1
-
-    while k < N_epochs:
-        # Outside
-        while t < N_samples:
-            if time_array[t] >= starts[k]:
-                break
-            t += 1
-
-        while i < N_target:
-            if time_target_array[i] >= starts[k]:
-                break
-            i += 1
-
-        if time_array[t] <= ends[k]:
-            start_t[k, 0] = t
-
-        if time_target_array[i] <= ends[k]:
-            start_t[k, 1] = i
-
-        # Inside
-        while t < N_samples:
-            if time_array[t] > ends[k]:
-                break
-            else:
-                count[k, 0] += 1
-            t += 1
-
-        while i < N_target:
-            if time_target_array[i] > ends[k]:
-                break
-            else:
-                count[k, 1] += 1
-            i += 1
-
-        k += 1
-
-        if k == N_epochs:
-            break
-        if t == N_samples:
-            break
-        if i == N_target:
-            break
-
-    new_data_array = np.full(
-        (np.sum(windowsize) + 1, np.sum(count[:, 1]), *data_array.shape[1:]), np.nan
-    )
-
-    if np.any((count[:, 0] * count[:, 1]) > 0):
-        for k in prange(N_epochs):
-            if count[k, 0] > 0 and count[k, 1] > 0:
-                t = start_t[k, 0]
-                i = start_t[k, 1]
-                maxt = t + count[k, 0]
-                maxi = i + count[k, 1]
-                cnt_i = np.sum(count[0:k, 1])
-
-                while i < maxi:
-                    interval = abs(time_array[t] - time_target_array[i])
-                    t_pos = t
-                    t += 1
-                    while t < maxt:
-                        new_interval = abs(time_array[t] - time_target_array[i])
-                        if new_interval > interval:
-                            break
-                        else:
-                            interval = new_interval
-                            t_pos = t
-                            t += 1
-
-                    left = np.minimum(windowsize[0], t_pos - start_t[k, 0])
-                    right = np.minimum(windowsize[1], maxt - t_pos - 1)
-                    center = windowsize[0] + 1
-                    new_data_array[center - left - 1 : center + right, cnt_i] = (
-                        data_array[t_pos - left : t_pos + right + 1]
-                    )
-
-                    t -= 1
-                    i += 1
-                    cnt_i += 1
-
-    return new_data_array
-
-
 ################################
 # IntervalSet functions
 ################################
